prepare

+++ front/svelte/transaction/transaction-entry.svelte
@@ -1,17 +1,15 @@
 <div class="entry">
-  <nav class="navbar navbar-expand-lg navbar-light bg-light">
-    <div class="container-fluid">
-      <h5 class="entry-title">取引文書(見積/請求/取引情報)</h5>
-      {#if transaction.no}
-      <span>管理番号:&nbsp;{transaction.no}</span>
-      {:else}
-      <span>新規</span>
-      {/if}
-    </div> 
-  </nav>
+  <div class="page-title d-flex justify-content-between">
+    <h1>取引文書(見積/請求/取引情報)</h1>
+    {#if transaction.no}
+    <span>管理番号:&nbsp;{transaction.no}</span>
+    {:else}
+    <span>新規</span>
+    {/if}
+  </div> 
   <div class="row full-height fontsize-12pt">
-    <div class="entry-content">
-      <div class="entry-body">
+    <div class="content">
+      <div class="body">
         <FormError
         	messages={errorMessages}></FormError>
         <TransactionInfo
@@ -21,12 +19,12 @@
           bind:transaction={transaction}
           bind:status={status}></TransactionInfo>
       </div>
-      <div class="entry-footer">
+      <div class="footer">
         <button type="button" class="btn btn-secondary" disabled={disabled}
           on:click={back}>もどる</button>
         {#if ( transaction && transaction.id && transaction.id > 0 )}
         <button type="button" class="btn btn-danger" disabled={disabled}
-          on:click={delete_}
+          on:click={deleteTransaction}
           id="delete-button">削除</button>
         <button type="button" class="btn btn-info" disabled={disabled}
           on:click={() => {
@@ -39,10 +37,24 @@
         <button type="button" class="btn btn-primary" disabled={disabled}
           on:click={save}
           id="save-button">保存</button>
+        {#if ( transaction && transaction.id && transaction.id > 0 )}
+        {#if ( transaction.kind.book.form )}
+        <a href="/forms/transaction/{transaction.kind.book.form}/{transaction.id}"
+        	class="btn btn-info" target="_blank" disabled={disabled}>
+          {transaction.kind.label}書作成
+        </a>
+        {/if}
+        {/if}
       </div>
     </div>
   </div>
 </div>
+<OkModal
+  bind:this={modal}
+  title={title}
+  description={description}
+  on:answer={operation}
+  ></OkModal>
 <script>
 import axios from 'axios';
 import {numeric, formatDate} from '../../../libs/utils.js';
@@ -50,6 +62,7 @@ import {onMount, beforeUpdate, afterUpdate, createEventDispatcher} from 'svelte'
 const dispatch = createEventDispatcher();
 import TransactionInfo from './transaction-info.svelte';
 import FormError from '../common/form-error.svelte';
+import OkModal from '../common/ok-modal.svelte';
 import {currentTransaction} from '../../javascripts/current-record.js'
 import {bindFile} from '../../javascripts/document.js';
 
@@ -57,10 +70,14 @@ export	let	status;
 export	let transaction;
 export	let users;
 
-let update;
 let disabled = false;
 let errorMessages = [];
 let files;
+
+let modal;
+let title;
+let description;
+let operation = () => {};
 
 const create_transaction = async (_transaction) => {
   let result = await axios.post('/api/transaction', _transaction);
@@ -74,10 +91,39 @@ const update_transaction = async (_transaction) => {
   console.log(result);
   return	(result);
 }
-const delete_transaction = async (transaction) => {
-  console.log('delete_transaction', transaction);
-  let result = await axios.delete(`/api/transaction/${transaction.id}`);
-  console.log(result);
+
+const deleteTransaction = (event) => {
+  console.log('deleteTransaction', transaction);
+  title = '取引の削除';
+  description = `
+<table style="font-size:12px;">
+  <tbody>
+    <tr>
+			<td>相手先</td><td>${transaction.customerName}</td>
+		</tr>
+    <tr>
+			<td>件名</td><td>${transaction.subject}</td>
+		</tr>
+    <tr>
+			<td>担当</td><td>${transaction.handleUser.member.tradingName}</td>
+    </tr>
+  </tbody>
+</table>
+`;
+  operation = doDelete;
+  modal.show();
+}
+
+const doDelete = async (event) => {
+  if	( event.detail )	{
+  	try {
+      let result = await axios.delete(`/api/transaction/${transaction.id}`);
+  		console.log(result);
+    	back();
+  	} catch (e) {
+	    console.log(e);
+  	}
+  }
 }
 
 const save = () => {
@@ -94,8 +140,8 @@ const save = () => {
     transaction.tax = numeric(transaction.tax);
   }
   transaction.taxClass = parseInt(transaction.taxClass);
-  console.log('kind', transaction.kind);
-  if  ( (!transaction.kind) || (transaction.kind == 0) )	{
+  console.log('kind', transaction.kindId);
+  if  ( (!transaction.kindId) || (transaction.kindId == 0) )	{
     ok = false;
     errorMessages.push('種別を入力してください。');
   }
@@ -117,21 +163,21 @@ const save = () => {
     	}
     	pr.then((result) => {
       	console.log('result', result);
-      	update = true;
       	if  ( !result.data.code ) {
-        	transaction = result.data.transaction;
-          bindFile(files,transaction.documentId);
-      	}
-				if	( create )	{
-        	window.history.replaceState(
-          	status, "", `/transaction/entry/${transaction.id}`);
-          axios(`/api/transaction/${transaction.id}`).then((result) => {
+        	let id = result.data.id;
+          let documentId = result.data.id;
+          bindFile(files,documentId);
+          axios.get(`/api/transaction/${id}`).then((result) => {
         		console.log('new load', result.data);
         		transaction = result.data.transaction;
         		currentTransaction.set(transaction);
+				    if	( create )	{
+        	    window.history.replaceState(
+          	    status, "", `/transaction/entry/${transaction.id}`);
+      	    } else {
+              currentTransaction.set(transaction);
+            }
           });
-      	} else {
-          currentTransaction.set(transaction);
         }
     	});
   	}
@@ -150,22 +196,11 @@ const	back = (event) => {
   window.history.back();
 }
 
+const makeVoucher = (event) => {
+
+}
+
 beforeUpdate(() => {
-  //console.log('transaction-entry beforeUpdate', transaction);
-  update = false;
 });
 
-const	delete_ = (event) => {
-  try {
-    console.log('delete');
-    delete_transaction(transaction).then(() => {
-      back();
-    });
-  }
-  catch(e) {
-    console.log(e);
-    // can't delete
-    //	TODO alert
-  }
-}
 </script>
